♻️ [Refactoring]: #441 lexer.rs をファサード化し Lexer 本体を core サブモジュールへ移設
- lexer/core.rs を新設し、Lexer 構造体と基本 API（new / position /
  cursor_position / buffer_capacity_for_tests / is_eof）を doc ごと移設
- フィールドは兄弟モジュールからの直接アクセス維持のため
  pub(in crate::lexer) に変更
- hex_value を byte_ops へ移設（pub(super) + 契約 doc、関数本体無変更）し、
  hex_string / name の import を super::byte_ops 経由に変更
- extern prelude の libcore との E0659 曖昧性回避のため self::core で再 export
- 挙動変更なし（cargo fmt --check / clippy -D warnings / test 1092 passed）
