+++ rust/crates/core/src/lexer/byte_ops.rs
@@ -4,6 +4,21 @@
 //! 本モジュールはトークン化ロジック（`hex_string` / `literal_string` 等）から
 //! 共通利用される計算ヘルパの集約先である。
 
+/// 16 進数字 1 バイト（`'0'-'9'` / `'a'-'f'` / `'A'-'F'`）を 0-15 のニブル値に変換する純関数。
+///
+/// # 契約
+/// - 呼び出し側で `is_ascii_hexdigit` を確認済みであることを前提とする。
+/// - 前提を満たす入力に対して panic しない（lexer 層の panic 不在契約に影響しない）。
+/// - 契約違反の入力（非 16 進数字）では結果は未定義相当であり、
+///   debug ビルドでは減算 overflow により panic しうる。
+pub(super) fn hex_value(b: u8) -> u8 {
+    match b {
+        b'0'..=b'9' => b - b'0',
+        b'a'..=b'f' => b - b'a' + 10,
+        _ => b - b'A' + 10,
+    }
+}
+
 /// PDF §7.3.4.3 16 進ペア合成: 上位 4bit と下位 4bit から 1 バイトを合成する純関数。
 ///
 /// # 契約
